requirements.txt
Flask==0.11.1
pyyaml
requests
urllib3
